requirement.txt
pyassimp == 4.1.3
progressbar2
pyopengl >= 3.1.0
opencv-python # == 4.2.0.34 but not supported
transforms3d
pillow
IPython
matplotlib
easydict
pyyaml
future
scipy
Cython
